Update text_field.dart add controlle and onchange
TextEditingController, used to control the content of the input box, monitor changes, obtain the current input text, etc.
onChanged: onChanged
This is a callback function. Whenever the user enters something in the input box, onChanged will be triggered. It receives the latest value entered by the user and can perform corresponding operations here, such as updating the status or performing form validation.
keyboardType: keyboardType
This attribute specifies what type of keyboard pops up when the user clicks the input box.