Update the user info page function
Check if the user name, email address, and phone number are unique when checking user information. Check if the user name, email address, and phone number input are in the correct format when the user information is modified. You can also choose to take a photo or use the gallery directly. Logout front-end and back-end connections can be made and can jump to the login page and change the user login state to false