Automatically fill in existing data and calendar reminders
Create an API that returns matching food information based on the food name entered when the user searches.
a search button is provided to allow the user to search the database for matching foods.
Automatically fill the food information returned by the backend into the input form, allowing users to see preset saving suggestions and information.
Logic for handling user input and calling the API to get suggestions
add a calendar reminder