Complete the shopping list
Created a dialog page for adding a shopping item that allows the user to enter the name, quantity and units of the item as well as input validation.
An Edit Shopping Item dialog page was created, similar to the Add Shopping Item page, which allows the user to modify the name, quantity, and units of an existing item.
Divide the shopping list into two parts: unpurchased items and purchased items. Added buttons to mark each item as purchased and unpurchased, added editing and deletion functions.