Synchronization of views and data after deleting an ingredient in the filter state and repairing the filter
Fixed view auto-update after deleting an ingredient: Retrieving the ingredient list after deleting an ingredient ensures that the page refreshes correctly.
And fix “The following assertion was thrown while finalizing the widget tree: Looking up a deactivated widget's ancestor is unsafe. this point the state of the widget's element tree is no longer stable.”, a bug in which the food type filter was thrown out of action was fixed.