Finish pop-up notification when adding an existing ingredient in shopping list
Duplicate Item Detection: Check if the same item already exists when adding shopping list items.
Conflict Handling Popup: If duplicate items are detected, a conflict popup will be displayed on the front-end to show the details of the duplicate items and provide two options: “Cancel” and “Add Anyway”.
Force Add Function: When the user selects “Add Anyway”, the system will call forceAddItem to force the data entered by the user to be added to the shopping list.
Data transfer correction: Ensure that the data entered by the user and not the existing data of the conflicting item is sent during the “Add Anyway” operation.