Implemented dynamic theme color switching
Added theme color selection to SettingsPage, allowing users to select different theme colors (e.g., “yellow-green theme” and “blue theme”) and store them in the back-end database to ensure that the next time the user logs in, he or she will be able to the next time the user logs in, ensuring that the previously selected theme can be restored.
Added handling of null responses to ensure that the application can use the default theme if the theme returned by the backend is null or incorrectly formatted.
Changed all page colors to use dynamic colors
Putting the logout on the setting page